requirements.txt
Wheel
pytest
pipreqs
flake8
black
Mongo
TwitchAPI